requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
supabase==1.0.3 